requirements.txt
pyyaml==6.0.3
psycopg2-binary==2.9.11
sqlalchemy==2.0.48
gitpython==3.1.46
